order business logic code refactoring

+++ adm/src/main/java/com/sh/adm/model/entity/OrderDetail.java
@@ -57,6 +57,7 @@ public class OrderDetail {
 
 
     public void detailTotalPrice() {
-        this.totalPrice = this.item.getPrice().multiply(BigDecimal.valueOf(getQuantity()));
+        this.totalPrice = (BigDecimal) this.item.getPrice().multiply(BigDecimal.valueOf(getQuantity()));
     }
+
 }
